menu api consolidation
